+++ raphodo/utilities.py
@@ -43,6 +43,7 @@ import pkg_resources
 
 import arrow
 import psutil
+from PyQt5.QtCore import QSize
 
 import raphodo.__about__ as __about__
 
@@ -930,3 +931,11 @@ def flexible_date_time_parser(dt_string: str) -> Tuple[datetime, str]:
     # no idea how to handle this properly -- so ignore for now!
 
     return datetime.strptime(dte, fs), fs
+
+def image_large_enough_fdo(size: QSize) -> bool:
+    """
+    :param size: image size
+    :return: True if image is large enough to meet the FreeDesktop.org
+     specs for a large thumbnail
+    """
+    return size.width() >= 256 or size.height() >= 256